Write compelling meta description
The meta description is limited to 100-150 characters on major search engines when shown after a search query. This description attempts to use as many keywords related to this page as possible while remaining engaging and within that character limit. Anymore would be cut off when displayed on search engine sites.