Provide keywords for SEO
While keywords are not as important to SEO as they used to be, they still serve as a valuable tool within SEO. They describe keywords that a search user may input to the search engine. If a search user types any of the keywords we have present, the search user is more likely to get referred to our page. However, it is only recommended that the keywords listed be related to your content; otherwise, search engines could penalize your page and rank it lower since it does not represent what a user is actually searching for.