Define content type
This metadata is important because it allows for the browser to properly understand your site. While the `text/html` is appended to the html document, it does not hurt to be explicit. Furthermore, the charset allows the browser to know what type of character encoding to use for your site. Using UTF-8 allows for many more characters than the original ASCII values.